+++ client/src/components/NavBar.jsx
@@ -1,4 +1,4 @@
-import { Link } from "react-router-dom";
+import { Link, NavLink } from "react-router-dom";
 import { useAuth } from "../hooks/useAuth.jsx";
 import "./NavBar.css";
 
@@ -7,17 +7,19 @@ export default function NavBar() {
 
   return (
     <nav className="navbar">
-      <Link to="/" className="nav-link">Search</Link>
-      {user ? (
-        <>
-          <Link to="/my-recipes" className="nav-link">My Recipes</Link>
-          <button className="nav-button" onClick={signOut}>Log Out</button>
-        </>
-      ) : (
-        <>
-          <Link to="/login" className="nav-link">Log In</Link>
-          <Link to="/signup" className="nav-link">Sign Up</Link>
-        </>
+      <div className="navbar-tabs">
+        <NavLink to="/" end className={({ isActive }) => `nav-tab${isActive ? " nav-tab--active" : ""}`}>
+          The Nutritionist
+        </NavLink>
+        <NavLink to="/herb-search" className={({ isActive }) => `nav-tab${isActive ? " nav-tab--active" : ""}`}>
+          Herb Search
+        </NavLink>
+      </div>
+      {user && (
+        <div className="navbar-aux">
+          <Link to="/my-recipes" className="nav-aux-link">My Recipes</Link>
+          <button className="nav-aux-button" onClick={signOut}>Log Out</button>
+        </div>
       )}
     </nav>
   );